code review over parser
