+++ bclib/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "3.14.2"
+__version__ = "3.15"
