+++ bclib/parser/answer/__init__.py
@@ -1,3 +1,4 @@
 from bclib.parser.answer.answer import Answer
 from bclib.parser.answer.user_action_types import UserActionTypes
 from bclib.parser.answer.user_action import UserAction
+from bclib.parser.answer.file_user_action import FileUserAction
