+++ bclib/parser/answer/user_action.py
@@ -1,14 +1,15 @@
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from bclib.parser.answer.answer import Answer
-from .user_action_types import UserActionTypes
+from ..answer.user_action_types import UserActionTypes
+from ..answer.file_user_action import FileUserAction
 
 
 class UserAction:
     """This class will be updated in next versions."""
 
-    def __init__(self, prp_id: int, action: UserActionTypes, prp_value_id: int, internal_prp_value_id:int, value_id: int, value: any, datatype: str,
-                 multi: bool, part: int, answer:'Answer'):
+    def __init__(self, prp_id: int, action: UserActionTypes, prp_value_id: int, internal_prp_value_id: int, value_id: int, value: any, datatype: str,
+                 multi: bool, part: int, answer: 'Answer'):
         self.prp_id = prp_id
         self.action = action
         self.prp_value_id = prp_value_id
@@ -17,11 +18,11 @@ class UserAction:
         self.part = part
         self.datatype = datatype
         self.multi = multi
-        self.answer:'Answer' = answer
-        self.internal_prp_value_id:int = internal_prp_value_id
-    
+        self.answer: 'Answer' = answer
+        self.internal_prp_value_id: int = internal_prp_value_id
+
     def as_tuple(self) -> tuple:
-        return (self.prp_id, self.action.value, self.prp_value_id,self.internal_prp_value_id, self.value_id,
+        return (self.prp_id, self.action.value, self.prp_value_id, self.internal_prp_value_id, self.value_id,
                 self.value, self.part, self.datatype, self.multi, self.answer)
 
     def as_dict(self) -> dict:
@@ -42,44 +43,14 @@ class UserAction:
         is_file = False
         if isinstance(self.value, dict):
             list(self.value.keys()).sort()
-            is_file = list(self.value.keys()) == ["content", "name", "size", "type"]
+            is_file = list(self.value.keys()) == [
+                "content", "name", "size", "type"]
         return is_file
-    
-    def as_file_content(self):
+
+    def as_file_content(self) -> 'FileUserAction':
         if self.is_file_content():
             value = self.value
             return FileUserAction(self.prp_id, self.prp_value_id, self.value_id, value["name"], value["type"], value["size"], value["content"])
-
-    def __str__(self) -> str:
-        return str(self.as_tuple())
-
-    def __repr__(self) -> str:
-        return str(self.as_dict())
-
-class FileUserAction:
-
-    def __init__(self, prp_id:"int", prp_value_id:"int", value_id:"int", name:"str", type:"str", size:"int", content:"str") -> None:
-        self.prp_id = prp_id
-        self.prp_value_id = prp_value_id
-        self.value_id = value_id
-        self.name = name
-        self.type = type
-        self.size = size
-        self.content = content
-
-    def as_tuple(self) -> tuple:
-        return (self.prp_id, self.prp_value_id, self.value_id, self.name, self.type, self.size, self.content)
-
-    def as_dict(self) -> dict:
-        return {
-            "prp_id": self.prp_id,
-            "prp_value_id": self.prp_value_id,
-            "value_id": self.value_id,
-            "name": self.name,
-            "type": self.type,
-            "size": self.size,
-            "content": self.content
-        }
 
     def __str__(self) -> str:
         return str(self.as_tuple())
